+++ src/main/java/net/minecraft/src/NetClientHandler.java
@@ -3,19 +3,16 @@ package net.minecraft.src;
 import java.io.ByteArrayInputStream;
 import java.io.DataInputStream;
 import java.io.IOException;
-import java.math.BigInteger;
-import java.security.PrivateKey;
-import java.security.PublicKey;
+import me.ayunami2000.ayuncraft.java.security.PublicKey;
 import java.util.*;
 
 import me.ayunami2000.ayuncraft.CryptManager;
 import net.lax1dude.eaglercraft.DefaultSkinRenderer;
 import net.lax1dude.eaglercraft.EaglerAdapter;
-import net.lax1dude.eaglercraft.EaglerProfile;
 import net.lax1dude.eaglercraft.WebsocketNetworkManager;
 import net.minecraft.client.Minecraft;
 
-import javax.crypto.SecretKey;
+import me.ayunami2000.ayuncraft.javax.crypto.SecretKey;
 
 public class NetClientHandler extends NetHandler {
