--- src/teavm/java/java/security/PrivateKey.java
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 1996, 2013, Oracle and/or its affiliates. All rights reserved.
- * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- *
- */
-
-package java.security;
-
-public interface PrivateKey extends Key, javax.security.auth.Destroyable {
-}
