--- src/teavm/java/javax/crypto/SecretKey.java
@@ -1,6 +1,0 @@
-package javax.crypto;
-
-import java.security.Key;
-
-public interface SecretKey extends Key {
-}
