+++ src/teavm/java/me/ayunami2000/ayuncraft/java/security/PrivateKey.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 1996, 2013, Oracle and/or its affiliates. All rights reserved.
+ * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package me.ayunami2000.ayuncraft.java.security;
+
+import me.ayunami2000.ayuncraft.javax.security.auth.Destroyable;
+
+public interface PrivateKey extends Key, Destroyable {
+}
