+++ src/teavm/java/me/ayunami2000/ayuncraft/java/security/PublicKey.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 1996, 2013, Oracle and/or its affiliates. All rights reserved.
+ * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package me.ayunami2000.ayuncraft.java.security;
+
+public interface PublicKey extends Key {
+}
