+++ src/teavm/java/me/ayunami2000/ayuncraft/java/security/spec/KeySpec.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 1997, 2013, Oracle and/or its affiliates. All rights reserved.
+ * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package me.ayunami2000.ayuncraft.java.security.spec;
+
+public interface KeySpec { }
