+++ src/teavm/java/me/ayunami2000/ayuncraft/javax/crypto/SecretKey.java
@@ -1,0 +1,6 @@
+package me.ayunami2000.ayuncraft.javax.crypto;
+
+import me.ayunami2000.ayuncraft.java.security.Key;
+
+public interface SecretKey extends Key {
+}
