+++ src/teavm/java/me/ayunami2000/ayuncraft/javax/security/auth/DestroyFailedException.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 1999, 2013, Oracle and/or its affiliates. All rights reserved.
+ * ORACLE PROPRIETARY/CONFIDENTIAL. Use is subject to license terms.
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ *
+ */
+
+package me.ayunami2000.ayuncraft.javax.security.auth;
+
+public class DestroyFailedException extends Exception {
+
+    private static final long serialVersionUID = -7790152857282749162L;
+
+    public DestroyFailedException() {
+        super();
+    }
+
+    public DestroyFailedException(String msg) {
+        super(msg);
+    }
+}
